refactor: move cursor details to own page
